test(recorded): share public api embedding fixture
